+++ packages/backend/src/index.ts
@@ -2,6 +2,7 @@
 // eslint-disable-next-line import/order
 import otelSdk from './otel'; // must be imported first
 
+import fs from 'fs';
 import { LightdashMode, SessionUser } from '@lightdash/common';
 import * as Sentry from '@sentry/node';
 import * as Tracing from '@sentry/tracing';
@@ -12,6 +13,7 @@ import connectSessionKnex from 'connect-session-knex';
 import cookieParser from 'cookie-parser';
 import express, { NextFunction, Request, Response } from 'express';
 import expressSession from 'express-session';
+import expressStaticGzip from 'express-static-gzip';
 import passport from 'passport';
 import refresh from 'passport-oauth2-refresh';
 import path from 'path';
@@ -169,9 +171,11 @@ if (
 // frontend assets - immutable because vite appends hash to filenames
 app.use(
     '/assets',
-    express.static(path.join(__dirname, '../../frontend/build/assets'), {
-        immutable: true,
-        maxAge: '1y',
+    expressStaticGzip(path.join(__dirname, '../../frontend/build/assets'), {
+        serveStatic: {
+            immutable: true,
+            maxAge: '1y',
+        },
     }),
 );