+++ packages/backend/src/scheduler/SchedulerTask.ts
@@ -1,29 +1,30 @@
 import {
     assertUnreachable,
     CompileProjectPayload,
+    CreateSchedulerAndTargets,
     CreateSchedulerLog,
+    CreateSchedulerTarget,
     DownloadCsvPayload,
     EmailNotificationPayload,
     getHumanReadableCronExpression,
     getRequestMethod,
+    getSchedulerUuid,
     GsheetsNotificationPayload,
     isChartValidationError,
+    isCreateScheduler,
+    isCreateSchedulerSlackTarget,
     isDashboardChartTileType,
     isDashboardValidationError,
-    isEmailTarget,
     isSchedulerCsvOptions,
     isSchedulerGsheetsOptions,
     isSchedulerImageOptions,
-    isSlackTarget,
     LightdashPage,
     NotificationPayloadBase,
     ScheduledDeliveryPayload,
-    Scheduler,
-    SchedulerEmailTarget,
+    SchedulerAndTargets,
     SchedulerFormat,
     SchedulerJobStatus,
     SchedulerLog,
-    SchedulerSlackTarget,
     SlackNotificationPayload,
     UploadMetricGsheetPayload,
     ValidateProjectPayload,
@@ -102,7 +103,7 @@ const getChartOrDashboard = async (
 };
 
 export const getNotificationPageData = async (
-    scheduler: Scheduler,
+    scheduler: CreateSchedulerAndTargets,
     jobId: string,
 ): Promise<NotificationPayloadBase['page']> => {
     const {
@@ -239,8 +240,13 @@ export const sendSlackNotification = async (
     jobId: string,
     notification: SlackNotificationPayload,
 ) => {
-    const { schedulerUuid, schedulerSlackTargetUuid, scheduledTime } =
-        notification;
+    const {
+        schedulerUuid,
+        schedulerSlackTargetUuid,
+        channel,
+        scheduledTime,
+        scheduler,
+    } = notification;
     analytics.track({
         event: 'scheduler_notification_job.started',
         anonymousId: LightdashAnalytics.anonymousId,
@@ -249,6 +255,7 @@ export const sendSlackNotification = async (
             schedulerId: schedulerUuid,
             schedulerTargetId: schedulerSlackTargetUuid,
             type: 'slack',
+            sendNow: schedulerUuid === undefined,
         },
     });
 
@@ -257,23 +264,8 @@ export const sendSlackNotification = async (
             throw new Error('Slack app is not configured');
         }
 
-        const scheduler =
-            await schedulerService.schedulerModel.getSchedulerAndTargets(
-                schedulerUuid,
-            );
-        const { format, savedChartUuid, dashboardUuid, name, targets, cron } =
-            scheduler;
+        const { format, savedChartUuid, dashboardUuid, name, cron } = scheduler;
 
-        const target = targets
-            .filter(isSlackTarget)
-            .find(
-                (t) => t.schedulerSlackTargetUuid === schedulerSlackTargetUuid,
-            );
-
-        if (!target) {
-            throw new Error('Slack destination not found');
-        }
-        const { channel } = target;
         await schedulerService.logSchedulerJob({
             task: 'sendSlackNotification',
             schedulerUuid,
@@ -306,11 +298,14 @@ export const sendSlackNotification = async (
                 details.description ? ` - ${details.description}` : ''
             }`,
             ctaUrl: url,
-            footerMarkdown: `This is a <${url}?scheduler_uuid=${schedulerUuid}|scheduled delivery> ${getHumanReadableCronExpression(
-                cron,
-            )} from Lightdash\n${
-                s3Service.getExpirationWarning()?.slack || ''
-            }`,
+            footerMarkdown:
+                schedulerUuid !== undefined
+                    ? `This is a <${url}?scheduler_uuid=${schedulerUuid}|scheduled delivery> ${getHumanReadableCronExpression(
+                          cron,
+                      )} from Lightdash\n${
+                          s3Service.getExpirationWarning()?.slack || ''
+                      }`
+                    : `This is a scheduled delivery triggered manually from Lightdash`,
         };
 
         if (format === SchedulerFormat.IMAGE) {
@@ -370,6 +365,7 @@ export const sendSlackNotification = async (
                 format,
                 resourceType:
                     pageType === LightdashPage.CHART ? 'chart' : 'dashboard',
+                sendNow: schedulerUuid === undefined,
             },
         });
         await schedulerService.logSchedulerJob({
@@ -393,6 +389,7 @@ export const sendSlackNotification = async (
                 schedulerId: schedulerUuid,
                 schedulerTargetId: schedulerSlackTargetUuid,
                 type: 'slack',
+                sendNow: schedulerUuid === undefined,
             },
         });
         await schedulerService.logSchedulerJob({
@@ -728,8 +725,13 @@ export const sendEmailNotification = async (
     jobId: string,
     notification: EmailNotificationPayload,
 ) => {
-    const { schedulerUuid, schedulerEmailTargetUuid, scheduledTime } =
-        notification;
+    const {
+        schedulerUuid,
+        schedulerEmailTargetUuid,
+        recipient,
+        scheduledTime,
+        scheduler,
+    } = notification;
 
     analytics.track({
         event: 'scheduler_notification_job.started',
@@ -739,27 +741,13 @@ export const sendEmailNotification = async (
             schedulerId: schedulerUuid,
             schedulerTargetId: schedulerEmailTargetUuid,
             type: 'email',
+            sendNow: schedulerUuid === undefined,
         },
     });
 
     try {
-        const scheduler =
-            await schedulerService.schedulerModel.getSchedulerAndTargets(
-                schedulerUuid,
-            );
-        const { format, savedChartUuid, dashboardUuid, name, targets } =
-            scheduler;
+        const { format, savedChartUuid, dashboardUuid, name } = scheduler;
 
-        const target = targets
-            .filter(isEmailTarget)
-            .find(
-                (t) => t.schedulerEmailTargetUuid === schedulerEmailTargetUuid,
-            );
-
-        if (!target) {
-            throw new Error('Email destination not found');
-        }
-        const { recipient } = target;
         await schedulerService.logSchedulerJob({
             task: 'sendEmailNotification',
             schedulerUuid,
@@ -843,6 +831,7 @@ export const sendEmailNotification = async (
                 withPdf: pdfFile !== undefined,
                 resourceType:
                     pageType === LightdashPage.CHART ? 'chart' : 'dashboard',
+                sendNow: schedulerUuid === undefined,
             },
         });
         await schedulerService.logSchedulerJob({
@@ -866,6 +855,7 @@ export const sendEmailNotification = async (
                 schedulerId: schedulerUuid,
                 schedulerTargetId: schedulerEmailTargetUuid,
                 type: 'email',
+                sendNow: schedulerUuid === undefined,
             },
         });
         await schedulerService.logSchedulerJob({
@@ -897,6 +887,7 @@ export const uploadGsheets = async (
             schedulerId: schedulerUuid,
             schedulerTargetId: undefined,
             type: 'gsheets',
+            sendNow: schedulerUuid === undefined,
         },
     });
 
@@ -1038,6 +1029,7 @@ export const uploadGsheets = async (
                 type: 'gsheets',
                 format,
                 resourceType: savedChartUuid ? 'chart' : 'dashboard',
+                sendNow: schedulerUuid === undefined,
             },
         });
         await schedulerService.logSchedulerJob({
@@ -1060,6 +1052,7 @@ export const uploadGsheets = async (
                 schedulerId: schedulerUuid,
                 schedulerTargetId: undefined,
                 type: 'gsheets',
+                sendNow: schedulerUuid === undefined,
             },
         });
         await schedulerService.logSchedulerJob({
@@ -1079,9 +1072,9 @@ export const uploadGsheets = async (
 
 const logScheduledTarget = async (
     format: SchedulerFormat,
-    target: SchedulerSlackTarget | SchedulerEmailTarget | undefined,
+    target: CreateSchedulerTarget | undefined,
     targetJobId: string,
-    schedulerUuid: string,
+    schedulerUuid: string | undefined,
     jobId: string,
     scheduledTime: Date,
 ) => {
@@ -1106,7 +1099,7 @@ const logScheduledTarget = async (
         CreateSchedulerLog,
         'task' | 'target' | 'targetType'
     > => {
-        if (isSlackTarget(target)) {
+        if (isCreateSchedulerSlackTarget(target)) {
             return {
                 task: 'sendSlackNotification',
                 target: target.channel,
@@ -1135,15 +1128,25 @@ const logScheduledTarget = async (
 export const handleScheduledDelivery = async (
     jobId: string,
     scheduledTime: Date,
-    { schedulerUuid }: ScheduledDeliveryPayload,
+    schedulerPayload: ScheduledDeliveryPayload,
 ) => {
+    const schedulerUuid = getSchedulerUuid(schedulerPayload);
+
     try {
+        const scheduler: SchedulerAndTargets | CreateSchedulerAndTargets =
+            isCreateScheduler(schedulerPayload)
+                ? schedulerPayload
+                : await schedulerService.schedulerModel.getSchedulerAndTargets(
+                      schedulerPayload.schedulerUuid,
+                  );
+
         analytics.track({
             event: 'scheduler_job.started',
             anonymousId: LightdashAnalytics.anonymousId,
             properties: {
                 jobId,
                 schedulerId: schedulerUuid,
+                sendNow: schedulerUuid === undefined,
             },
         });
         await schedulerService.logSchedulerJob({
@@ -1154,11 +1157,6 @@ export const handleScheduledDelivery = async (
             scheduledTime,
             status: SchedulerJobStatus.STARTED,
         });
-
-        const scheduler =
-            await schedulerService.schedulerModel.getSchedulerAndTargets(
-                schedulerUuid,
-            );
 
         const page =
             scheduler.format === SchedulerFormat.GSHEETS