+++ packages/common/src/types/scheduler.ts
@@ -105,6 +105,19 @@ export type SchedulerEmailTarget = {
 export type CreateSchedulerTarget =
     | Pick<SchedulerSlackTarget, 'channel'>
     | Pick<SchedulerEmailTarget, 'recipient'>;
+
+export const getSchedulerTargetUuid = (
+    target: SchedulerSlackTarget | SchedulerEmailTarget | CreateSchedulerTarget,
+): string | undefined => {
+    if ('schedulerSlackTargetUuid' in target) {
+        return target.schedulerSlackTargetUuid;
+    }
+    if ('schedulerEmailTargetUuid' in target) {
+        return target.schedulerEmailTargetUuid;
+    }
+    return undefined;
+};
+
 export type UpdateSchedulerSlackTarget = Pick<
     SchedulerSlackTarget,
     'schedulerSlackTargetUuid' | 'channel'
@@ -217,10 +230,21 @@ export type ApiSchedulerLogsResponse = {
     status: 'ok';
     results: SchedulerWithLogs;
 };
+export type ApiTestSchedulerResponse = {
+    status: 'ok';
+};
 
 // Scheduler task types
-
-export type ScheduledDeliveryPayload = { schedulerUuid: string };
+export type ScheduledDeliveryPayload =
+    | { schedulerUuid: string }
+    | CreateSchedulerAndTargets;
+export const isCreateScheduler = (
+    data: ScheduledDeliveryPayload,
+): data is CreateSchedulerAndTargets => 'targets' in data;
+export const getSchedulerUuid = (
+    data: ScheduledDeliveryPayload,
+): string | undefined =>
+    isCreateScheduler(data) ? undefined : data.schedulerUuid;
 
 export enum LightdashPage {
     DASHBOARD = 'dashboard',
@@ -229,7 +253,7 @@ export enum LightdashPage {
 }
 
 export type NotificationPayloadBase = {
-    schedulerUuid: string;
+    schedulerUuid?: string;
     scheduledTime: Date;
     jobGroup: string;
     page: {
@@ -255,14 +279,17 @@ export type NotificationPayloadBase = {
         }[];
         pdfFile?: string;
     };
+    scheduler: CreateSchedulerAndTargets;
 };
 
 export type SlackNotificationPayload = NotificationPayloadBase & {
-    schedulerSlackTargetUuid: string;
+    schedulerSlackTargetUuid?: string;
+    channel: string;
 };
 
 export type EmailNotificationPayload = NotificationPayloadBase & {
-    schedulerEmailTargetUuid: string;
+    schedulerEmailTargetUuid?: string;
+    recipient: string;
 };
 
 export type GsheetsNotificationPayload = {