+++ packages/frontend/src/components/DashboardFilter/FilterConfiguration/TileFilterConfiguration.tsx
@@ -15,6 +15,7 @@ import {
     Flex,
     Stack,
     Text,
+    Tooltip,
     useMantineTheme,
 } from '@mantine/core';
 import { FC, useCallback, useMemo } from 'react';
@@ -93,7 +94,20 @@ const TileFilterConfiguration: FC<Props> = ({
     const tileTargetList = useMemo(() => {
         return sortedTileWithFilters.map(([tileUuid, filters], index) => {
             const tile = tiles.find((t) => t.uuid === tileUuid);
+
+            // tileConfig overrides the default filter state for a tile
+            // if it is a field, we use that field for the filter.
+            // If it is the empty string, the filter is disabled.
             const tileConfig = filterRule.tileTargets?.[tileUuid];
+
+            let selectedField;
+            if (tileConfig !== false) {
+                selectedField = tileConfig?.fieldId
+                    ? filters?.find(
+                          (f) => tileConfig?.fieldId === getFieldId(f),
+                      )
+                    : filters?.find((f) => matchFieldExact(f)(field));
+            }
 
             const isFilterAvailable =
                 filters?.some(matchFieldByType(field)) ?? false;
@@ -104,11 +118,6 @@ const TileFilterConfiguration: FC<Props> = ({
                     sortFieldsByMatch(matchFieldByTypeAndName, a, b),
                 )
                 .sort((a, b) => sortFieldsByMatch(matchFieldExact, a, b));
-
-            const fieldId = tileConfig?.fieldId;
-            const selectedFilter = filters?.find(
-                (f) => getFieldId(f) === fieldId,
-            );
 
             const tileWithoutTitle =
                 !tile?.properties.title || tile.properties.title.length === 0;
@@ -126,7 +135,8 @@ const TileFilterConfiguration: FC<Props> = ({
             return {
                 key: tileUuid + index,
                 label: tileLabel,
-                checked: isFilterAvailable && !!tileConfig,
+                checked: !!selectedField,
+                disabled: !isFilterAvailable,
                 tileUuid,
                 ...(tile &&
                     isDashboardChartTileType(tile) && {
@@ -134,7 +144,7 @@ const TileFilterConfiguration: FC<Props> = ({
                             tile.properties.lastVersionChartKind ?? undefined,
                     }),
                 sortedFilters,
-                selectedFilter,
+                selectedField,
             };
         });
     }, [filterRule, field, sortFieldsByMatch, sortedTileWithFilters, tiles]);
@@ -175,33 +185,44 @@ const TileFilterConfiguration: FC<Props> = ({
             <Stack spacing="md">
                 {tileTargetList.map((value) => (
                     <Box key={value.key}>
-                        <Checkbox
-                            size="xs"
-                            fw={500}
-                            label={
-                                <Flex align="center" gap="xxs">
-                                    <MantineIcon
-                                        color="blue.8"
-                                        icon={getChartIcon(value.tileChartKind)}
-                                    />
-                                    {value.label}
-                                </Flex>
-                            }
-                            styles={{
-                                label: {
-                                    paddingLeft: theme.spacing.xs,
-                                },
-                            }}
-                            checked={value.checked}
-                            onChange={(event) => {
-                                onChange(
-                                    event.currentTarget.checked
-                                        ? FilterActions.ADD
-                                        : FilterActions.REMOVE,
-                                    value.tileUuid,
-                                );
-                            }}
-                        />
+                        <Tooltip
+                            label="No fields matching filter type"
+                            position="left"
+                            disabled={!value.disabled}
+                        >
+                            <Box>
+                                <Checkbox
+                                    size="xs"
+                                    fw={500}
+                                    disabled={value.disabled}
+                                    label={
+                                        <Flex align="center" gap="xxs">
+                                            <MantineIcon
+                                                color="blue.8"
+                                                icon={getChartIcon(
+                                                    value.tileChartKind,
+                                                )}
+                                            />
+                                            {value.label}
+                                        </Flex>
+                                    }
+                                    styles={{
+                                        label: {
+                                            paddingLeft: theme.spacing.xs,
+                                        },
+                                    }}
+                                    checked={value.checked}
+                                    onChange={(event) => {
+                                        onChange(
+                                            event.currentTarget.checked
+                                                ? FilterActions.ADD
+                                                : FilterActions.REMOVE,
+                                            value.tileUuid,
+                                        );
+                                    }}
+                                />
+                            </Box>
+                        </Tooltip>
 
                         {value.sortedFilters && (
                             <Box
@@ -227,7 +248,7 @@ const TileFilterConfiguration: FC<Props> = ({
                                         },
                                     }}
                                     fields={value.sortedFilters}
-                                    activeField={value.selectedFilter}
+                                    activeField={value.selectedField}
                                     onChange={(newFilter) => {
                                         onChange(
                                             FilterActions.ADD,